Just fixed program that finds and prints the largest prime factor of the number 612852475143 @11

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -6,8 +6,8 @@
  */
 int main(void)
 {
-	long long num = 612852475143;
-	long long factor = 2;
+	long num = 612852475143;
+	long factor = 2;
 
 	while (num > 1)
 	{
@@ -20,7 +20,7 @@ int main(void)
 			factor++;
 		}
 	}
-	printf("%lld\n", factor);
+	printf("%li\n", factor);
 
 	return 0;
 }
